requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.12
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
Django==4.1.2
django-excel==0.0.10
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.8.0
lml==0.1.0
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pycparser==2.21
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
PyJWT==2.6.0
pylint==2.15.5
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.5
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
texttable==1.6.4
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
